refactor: play logic hooks
